Goodness-of-fit assessment of the fitted model.

gof() calculates some omnibus discrepancy measures and their Bayesian \(p\)-values for the fitted model using the posterior predictive check approach.

Usage

gof(fit, stats = c("Freeman_Tukey", "deviance"), cores = 1L, plot = TRUE)

Arguments

fit

An occumbFit object.

stats

The discrepancy statistics to be applied.

cores

The number of cores to use for parallelization.

plot

Logical, determine if draw scatter plots of the fit statistics.

Value

A list with the following named elements:

stats

The discrepancy statistics applied.

p_value

Bayesian \(p\)-value.

stats_obs

Discrepancy statistics for observed data.

stats_rep

Discrepancy statistics for repeated data.

Details

A discrepancy statistic of the fitted model is obtained using the procedure of posterior predictive checking. The following statistics are currently available:

Freeman-Tukey statistics (default)

\(T(\boldsymbol{y}, \boldsymbol{\theta}) = \sum_{i}\sum_{j}\sum_{k}\left(\sqrt{y_{ijk}} - \sqrt{E(y_{ijk} \mid \boldsymbol{\theta})}\right)^2\)

Deviance statistics

\(T(\boldsymbol{y}, \boldsymbol{\theta}) = -2 \log p(\boldsymbol{y} \mid \boldsymbol{\theta})\)

where \(\boldsymbol{y} = \{y_{ijk}\}\), \(\boldsymbol{\theta}\), \(E(y_{ijk} \mid \boldsymbol{\theta})\), and \(\log p(\boldsymbol{y} \mid \boldsymbol{\theta})\) are sequence read count data, parameters and latent variables of the model, expected value of the data conditional on \(\boldsymbol{\theta}\), and log-likelihood of the model, respectively. The Bayesian \(p\)-value is estimated as the probability that the value of the discrepancy statistics of replicated dataset is more extreme than that of the observed dataset. An extreme Bayesian \(p\)-value may indicate an inadequate model fit. See, e.g., Gelman et al. (2014), Kéry and Royle (2016), and Conn et al. (2018) for more details on the procedures for posterior predictive checking.

Computations can be run in parallel on multiple CPU cores where the cores argument controls the degree of parallelization.

References

P. B. Conn, D. S. Johnson, P. J. Williams, S. R. Melin and M. B. Hooten. (2018) A guide to Bayesian model checking for ecologists. Ecological Monographs 88:526--542. doi:10.1002/ecm.1314

A. Gelman, J. B. Carlin, H. S. Stern D. B. Dunson, A. Vehtari and D. B. Rubin (2013) Bayesian Data Analysis. 3rd edition. Chapman and Hall/CRC. http://www.stat.columbia.edu/~gelman/book/

M. Kéry and J. A. Royle (2016) Applied Hierarchical Modeling in Ecology --- Analysis of Distribution, Abundance and Species Richness in R and BUGS. Volume 1: Prelude and Static Models. Academic Press. https://www.mbr-pwrc.usgs.gov/pubanalysis/keryroylebook/

Examples

if (FALSE) {
# Generate the smallest random dataset (2 species * 2 sites * 2 reps)
I <- 2 # Number of species
J <- 2 # Number of sites
K <- 2 # Number of replicates
data <- occumbData(
    y = array(sample.int(I * J * K), dim = c(I, J, K)),
    spec_cov = list(cov1 = rnorm(I)),
    site_cov = list(cov2 = rnorm(J),
                    cov3 = factor(1:J)),
    repl_cov = list(cov4 = matrix(rnorm(J * K), J, K)))

# Fitting a null model
fit <- occumb(data = data)

# Goodness-of-fit assessment
gof_result <- gof(fit)
gof_result$p_value  # print p-value
}